+++ crates/lexgen/src/dfa.rs
@@ -109,23 +109,14 @@ impl<A> DFA<StateIdx, A> {
         self.states[next.0].predecessors.insert(state);
     }
 
-    pub fn add_range_transition(
-        &mut self,
-        state: StateIdx,
-        range_begin: u32,
-        range_end: u32,
-        next: StateIdx,
-    ) {
-        self.states[state.0].range_transitions.insert(
-            range_begin,
-            range_end,
-            next,
-            |_states_1, _states_2| {
-                panic!("Overlapping range transition in DFA::add_range_transition")
-            },
-        );
+    pub fn set_range_transitions(&mut self, state: StateIdx, range_map: RangeMap<StateIdx>) {
+        assert!(self.states[state.0].range_transitions.is_empty());
 
-        self.states[next.0].predecessors.insert(state);
+        for range in range_map.iter() {
+            self.states[range.value.0].predecessors.insert(state);
+        }
+
+        self.states[state.0].range_transitions = range_map;
     }
 
     pub fn set_any_transition(&mut self, state: StateIdx, next: StateIdx) {
@@ -188,7 +179,6 @@ impl<A> DFA<StateIdx, A> {
         } in other.states
         {
             let mut new_char_transitions: Map<char, StateIdx> = Default::default();
-            let mut new_range_transitions: RangeMap<StateIdx> = Default::default();
             let mut new_any_transition: Option<StateIdx> = None;
             let mut new_end_of_input_transition: Option<StateIdx> = None;
 
@@ -196,14 +186,8 @@ impl<A> DFA<StateIdx, A> {
                 new_char_transitions.insert(char, StateIdx(next.0 + n_current_states));
             }
 
-            for range in range_transitions.iter() {
-                new_range_transitions.insert(
-                    range.start,
-                    range.end,
-                    StateIdx(range.value.0 + n_current_states),
-                    |_states_1, _states_2| panic!("Overlapping range transition in DFA::add_dfa"),
-                );
-            }
+            let new_range_transitions =
+                range_transitions.map(|state_idx| StateIdx(state_idx.0 + n_current_states));
 
             if let Some(next) = any_transition {
                 new_any_transition = Some(StateIdx(next.0 + n_current_states));