+++ crates/lexgen/src/nfa.rs
@@ -111,18 +111,31 @@ impl<A> NFA<A> {
     pub fn add_range_transition(
         &mut self,
         state: StateIdx,
-        range_begin: char,
+        range_start: char,
         range_end: char,
         next: StateIdx,
     ) {
         let mut set: Set<StateIdx> = Default::default();
         set.insert(next);
         self.states[state.0].range_transitions.insert(
-            range_begin as u32,
+            range_start as u32,
             range_end as u32,
             set,
             |values_1, values_2| values_1.extend(values_2.into_iter()),
         );
+    }
+
+    pub fn add_range_transitions(&mut self, state: StateIdx, ranges: RangeMap<()>, next: StateIdx) {
+        let mut set: Set<StateIdx> = Default::default();
+        set.insert(next);
+
+        let ranges = ranges.map(|()| set.clone());
+
+        self.states[state.0]
+            .range_transitions
+            .insert_ranges(ranges.into_iter(), |values_1, values_2| {
+                values_1.extend(values_2.into_iter())
+            });
     }
 
     pub fn add_empty_transition(&mut self, state: StateIdx, next: StateIdx) {