+++ crates/lexgen/src/regex_to_nfa.rs
@@ -1,9 +1,8 @@
-use crate::ast::{CharOrRange, Regex, Var};
-use crate::builtin::BUILTIN_RANGES;
+use crate::ast::{Builtin, CharOrRange, Regex, Var};
+use crate::builtin::{BuiltinCharRange, BUILTIN_RANGES};
 use crate::collections::Map;
 use crate::nfa::{StateIdx, NFA};
-
-use std::convert::TryFrom;
+use crate::range_map::{Range, RangeMap};
 
 pub fn add_re<A>(
     nfa: &mut NFA<A>,
@@ -14,25 +13,22 @@ pub fn add_re<A>(
 ) {
     match re {
         Regex::Builtin(builtin_name) => {
-            let builtin = BUILTIN_RANGES.iter().find_map(|(name, builtin_)| {
-                if *name == builtin_name.0 {
-                    Some(*builtin_)
-                } else {
-                    None
-                }
-            });
+            let builtin = get_builtin_regex(builtin_name);
 
-            let builtin =
-                builtin.unwrap_or_else(|| panic!("Unknown builtin regex: {}", builtin_name.0));
+            let ranges: Vec<Range<()>> = builtin
+                .get_ranges()
+                .iter()
+                .copied()
+                .map(|(start, end)| Range {
+                    start,
+                    end,
+                    value: (),
+                })
+                .collect();
 
-            for (range_start, range_end) in builtin.get_ranges() {
-                nfa.add_range_transition(
-                    current,
-                    char::try_from(*range_start).unwrap(),
-                    char::try_from(*range_end).unwrap(),
-                    cont,
-                );
-            }
+            let map = RangeMap::from_non_overlapping_sorted_ranges(ranges);
+
+            nfa.add_range_transitions(current, map, cont);
         }
 
         Regex::Var(var) => {
@@ -122,5 +118,118 @@ pub fn add_re<A>(
         Regex::EndOfInput => {
             nfa.add_end_of_input_transition(current, cont);
         }
+
+        Regex::Diff(_, _) => {
+            let map = regex_to_range_map(bindings, re);
+            nfa.add_range_transitions(current, map, cont);
+        }
     }
 }
+
+fn get_builtin_regex(builtin: &Builtin) -> BuiltinCharRange {
+    BUILTIN_RANGES
+        .iter()
+        .find_map(|(name, builtin_)| {
+            if *name == builtin.0 {
+                Some(*builtin_)
+            } else {
+                None
+            }
+        })
+        .unwrap_or_else(|| panic!("Unknown builtin regex: {}", builtin.0))
+}
+
+fn regex_to_range_map(bindings: &Map<Var, Regex>, re: &Regex) -> RangeMap<()> {
+    match re {
+        Regex::Builtin(builtin) => {
+            let builtin = get_builtin_regex(builtin);
+            let ranges: Vec<Range<()>> = builtin
+                .get_ranges()
+                .iter()
+                .copied()
+                .map(|(start, end)| Range {
+                    start,
+                    end,
+                    value: (),
+                })
+                .collect();
+            RangeMap::from_non_overlapping_sorted_ranges(ranges)
+        }
+
+        Regex::Var(var) => {
+            let re = bindings
+                .get(var)
+                .unwrap_or_else(|| panic!("Unbound variable {:?}", var.0));
+
+            regex_to_range_map(bindings, re)
+        }
+
+        Regex::Char(char) => {
+            let mut map = RangeMap::new();
+            map.insert(*char as u32, *char as u32, (), merge_values);
+            map
+        }
+
+        Regex::String(_) => panic!("strings cannot be used in char sets (`#`)"),
+
+        Regex::CharSet(char_set) => {
+            let mut map = RangeMap::new();
+
+            // TODO: Quadratic behavior below, `RangeMap::insert` is O(number of ranges)
+            for char_or_range in char_set.0.iter() {
+                match char_or_range {
+                    CharOrRange::Char(char) => {
+                        map.insert(*char as u32, *char as u32, (), merge_values);
+                    }
+                    CharOrRange::Range(start, end) => {
+                        map.insert(*start as u32, *end as u32, (), merge_values);
+                    }
+                }
+            }
+
+            map
+        }
+
+        Regex::ZeroOrMore(_) => {
+            panic!("`*` cannot be used in char sets (`#`)");
+        }
+
+        Regex::OneOrMore(_) => {
+            panic!("`+` cannot be used in char sets (`#`)");
+        }
+
+        Regex::ZeroOrOne(_) => {
+            panic!("`?` cannot be used in char sets (`#`)");
+        }
+
+        Regex::Concat(_, _) => {
+            panic!("concatenation (`<re1> <re2>`) cannot be used in char sets (`#`)");
+        }
+
+        Regex::Or(re1, re2) => {
+            let mut map1 = regex_to_range_map(bindings, re1);
+            let map2 = regex_to_range_map(bindings, re2);
+
+            map1.insert_ranges(map2.into_iter(), merge_values);
+
+            map1
+        }
+
+        Regex::Any => {
+            let mut map = RangeMap::new();
+            map.insert(0, char::MAX as u32, (), merge_values);
+            map
+        }
+
+        Regex::EndOfInput => panic!("`$` cannot be used in char sets (`#`)"),
+
+        Regex::Diff(re1, re2) => {
+            let mut map1 = regex_to_range_map(bindings, re1);
+            let map2 = regex_to_range_map(bindings, re2);
+            map1.remove_ranges(&map2);
+            map1
+        }
+    }
+}
+
+fn merge_values(_val1: &mut (), _val2: ()) {}