+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+- New regex syntax `#` added for character set difference, e.g. `re1 # re2`
+  matches characters in `re1` that are not in `re2`. `re1` and `re2` need to be
+  "character sets", i.e. `*`, `+`, `?`, `"..."`, `$`, and concatenation are not
+  allowed.
+
 # 2021/11/30: 0.8.1
 
 New version published to fix broken README pages for lexgen and lexgen_util in